Move TrinoAwsProxyConfig to top-level package
