+++ trino-aws-proxy/src/main/java/io/trino/aws/proxy/server/TrinoAwsProxyConfig.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.trino.aws.proxy.server.rest;
+package io.trino.aws.proxy.server;
 
 import io.airlift.configuration.Config;
 import io.airlift.configuration.ConfigDescription;
