+++ trino-aws-proxy/src/test/java/io/trino/aws/proxy/server/testing/TestingS3ClientProvider.java
@@ -16,7 +16,7 @@ package io.trino.aws.proxy.server.testing;
 import com.google.inject.Inject;
 import com.google.inject.Provider;
 import io.airlift.http.server.testing.TestingHttpServer;
-import io.trino.aws.proxy.server.rest.TrinoAwsProxyConfig;
+import io.trino.aws.proxy.server.TrinoAwsProxyConfig;
 import io.trino.aws.proxy.server.testing.TestingUtil.ForTesting;
 import io.trino.aws.proxy.spi.credentials.Credential;
 import io.trino.aws.proxy.spi.credentials.Credentials;
